huggingface/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
gradio>=4.0.0
openai-whisper>=20231117
numpy>=1.24.0
soundfile>=0.12.0
librosa>=0.10.0
